Split session history event utilities
